+++ examples/tetris.rb
@@ -109,22 +109,23 @@ class Tetris
     menu('Game') {
       @pause_menu_item = check_menu_item('Pause') {
         enabled false
-        
-        on_clicked do
-          @game.paused = @pause_menu_item.checked?
-        end
+        checked <=> [@game, :paused]
       }
+      
       menu_item('Restart') {
         on_clicked do
           @game.restart!
         end
       }
+      
       separator_menu_item
+      
       menu_item('Exit') {
         on_clicked do
           exit(0)
         end
       }
+      
       quit_menu_item if OS.mac?
     }
     
@@ -134,6 +135,7 @@ class Tetris
           show_high_scores
         end
       }
+      
       menu_item('Clear High Scores') {
         on_clicked {
           @game.clear_high_scores!
@@ -147,11 +149,13 @@ class Tetris
           @game.instant_down_on_up = true
         end
       }
+      
       radio_menu_item('Rotate Right on Up Arrow') {
         on_clicked do
           @game.rotate_right_on_up = true
         end
       }
+      
       radio_menu_item('Rotate Left on Up Arrow') {
         checked true
 
@@ -169,6 +173,7 @@ class Tetris
           end
         }
       end
+      
       menu_item('About') {
         on_clicked do
           show_about_dialog