+++ lib/glimmer/libui/control_proxy/checkbox_proxy.rb
@@ -30,8 +30,7 @@ module Glimmer
       class CheckboxProxy < ControlProxy
         DEFAULT_TEXT = ''
       
-        def data_bind(property, model_binding)
-          super
+        def data_bind_write(property, model_binding)
           handle_listener('on_toggled') { model_binding.call(checked) } if property == 'checked'
         end