+++ lib/glimmer/libui/control_proxy/color_button_proxy.rb
@@ -114,8 +114,7 @@ module Glimmer
           super
         end
         
-        def data_bind(property, model_binding)
-          super
+        def data_bind_write(property, model_binding)
           handle_listener('on_changed') { model_binding.call(color) } if property == 'color'
         end
       end