+++ lib/glimmer/libui/control_proxy/date_time_picker_proxy.rb
@@ -65,8 +65,7 @@ module Glimmer
           super
         end
         
-        def data_bind(property, model_binding)
-          super
+        def data_bind_write(property, model_binding)
           handle_listener('on_changed') { model_binding.call(time) } if property == 'time'
         end
       end