+++ lib/glimmer/libui/control_proxy/menu_item_proxy/check_menu_item_proxy.rb
@@ -29,8 +29,7 @@ module Glimmer
         #
         # Follows the Proxy Design Pattern
         class CheckMenuItemProxy < MenuItemProxy
-          def data_bind(property, model_binding)
-            super
+          def data_bind_write(property, model_binding)
             handle_listener('on_clicked') { model_binding.call(checked) } if property == 'checked'
           end