+++ lib/glimmer/libui/control_proxy/menu_item_proxy/radio_menu_item_proxy.rb
@@ -65,8 +65,7 @@ module Glimmer
             end
           end
         
-          def data_bind(property, model_binding)
-            super
+          def data_bind_write(property, model_binding)
             handle_listener('on_clicked') { model_binding.call(checked) } if property == 'checked'
           end