+++ lib/glimmer/libui/data_bindable.rb
@@ -23,9 +23,25 @@ module Glimmer
   module LibUI
     # Parent controls and shapes who have children and add child post_initialize_child
     module DataBindable
-      # Data-binds model to update view.
-      # Subclasses can override to do inverse data-binding by observing view control for property changes and updating model binding accordingly
+      # Sets up read/write (bidirectional) data-binding
+      #
+      # classes are expected to implement `data_bind_write(property, model_binding)` to setup write data-binding
+      # by observing view property for changes and writing to model attribute via model binding accordingly
+      #
+      # classes can override data_bind_read to disable read data-binding in rare scenarios that might need it
+      #
+      # returns model attribute reading observer by default
       def data_bind(property, model_binding)
+        data_bind_read(property, model_binding).tap do
+          data_bind_write(property, model_binding) unless model_binding.binding_options[:read_only]
+        end
+      end
+      
+      # Sets up read data-binding (reading from model to update view)
+      #
+      # Default implementation observes model attribute for changes via model binding
+      # and updates view property accordingly
+      def data_bind_read(property, model_binding)
         model_attribute_observer = Glimmer::DataBinding::Observer.proc do
           new_value = model_binding.evaluate_property
           send("#{property}=", new_value) unless send(property) == new_value
@@ -33,6 +49,15 @@ module Glimmer
         model_attribute_observer.observe(model_binding)
         model_attribute_observer.call # initial update
         model_attribute_observer
+      end
+      
+      # Sets up write data-binding (writing to model from view)
+      #
+      # Has no implementation by default. Classes are expected
+      # to implement this method by observing view property
+      # for changes and writing them to model accordingly via model binding
+      def data_bind_write(property, model_binding)
+        # No Op by default
       end
     end
   end